Refactored Dialog.js: - Created a separate component, LegacyNonMarkdownRenderer, that is used inside Dialog if Admin wants to bypass the new ReactMarkdown renderer. This legacy solution still uses dangeresoulySetInnerHTML, which is what some users still want to do. - This legacy renderer is also **the default renderer** for the old Information plugin. We preserve the current functionality by hard coding this option. If you want to use ReactMarkdown, you should switch to the new plugin: InfoDialog. - Dialog.js has been moved inside the Dialog folder. Some other componenets that import Dialog are included in this commit where I've fixed the path. - DialogWindowPlugin now passes on all the options to Dialog. This way we it's a bit more future-proof as we won't have to modify it each time we which to send a new option to the Dialog.
